move login from home to login; create new home page, add images, button shop now, and welcome text
